rename track to public class, consolidate track and identify into public class
